bash: Add script ts - timestamp input
ts adds a timestamp to the beginning of each line of input.

This script is a wrapper around `date' command, all options to `ts' are
passd to `date' as is, with default date format changed to `%H:%M:%S'.

This script is intended to be a simple alternative to `ts' of
`moreutils' package.