Code: Compact pinned tabs and highlight modified ones
Those are especially useful for terminal tabs.

A terminal tab always has a modified state. When it's not pinned, a bit
dot is shown, it's much better visually to simply highlight it instead.
If a terminal tab is pinned, we can compact it, because chances are we
do not care about its title.